Neo4jService interface
