+++ BACKLOG.md
@@ -1,0 +1,8 @@
+### Backlog
+
+* IDEA: Add support for bolt driver
+* IDEA: Load adoc files with Cypher scripts in code blocks
+* IDEA: Create standard CSV and JSON load templates
+
+
+* DONE: Basic `Neo4jService` and test to create and query data
